Graphical Window; Intersect Circle
This program was my first experience with utilizing the graphical window in Python. The intention of the assignment was to draw a circle and an intersecting horizontal line based on user input and display red dots at the intersecting points. The x-axis values for both intersecting points are to be displayed as text items in the graphical window after the objects have been drawn.